% 
% TODO: Write documentation 
% 
\name{RDyNetNode}
\alias{RDyNetNode}
\title{
	RDyNetNode class
}
\description{
	RDyNetNode is a class defining labels, ids and all other attributes of nodes in a RDyNetGraph.      
}
\usage{
	RDyNetNode(node_tab)
}
\arguments{
	\item{node_tab}{
		a table includes labels, ids, and all other attributes of nodes. 
	}
}
\value{
	a RDyNetNode object
}
\details{
	Once a RDyNetNode object has been created, the attributes of nodes can be changed by referring to its member function SetNodeAttr.
}
\author{
	Xin Wang
}
\examples{
	# demo("demo", "IGIR")
}

